+++ src/parser/utils.rs
@@ -59,7 +59,9 @@ pub(crate) fn parse_string(string: Pair<Rule>) -> String {
             quoted[1..len - 1].to_string()
         }
         Rule::unquoted => string.as_str().to_string(),
-        Rule::currency => parse_string(string.into_inner().next().unwrap()),
+        Rule::currency | Rule::commodity_in_directive => {
+            parse_string(string.into_inner().next().unwrap())
+        }
         _ => string.as_str().to_string(),
     }
 }