+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog file for dinero-rs project, a command line application for managing fi
 ## [0.21.0] - xxx
 ### Added
 - Infer currency format from the journal file
+- ```isin``` is a valid property for commodities
 ### Changed
 - Continuos integration pipeline is now better. No more problems like what happened between releases 0.18 and 0.20.
 ### Fixed